requirement.txt
asgiref==3.3.1
certifi==2019.9.11
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
pipenv==2018.11.26
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
